ml_mini_project/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
firebase-admin==6.2.0
requests==2.31.0
numpy==1.26.0
scikit-learn==1.3.1
joblib==1.3.2
python-dotenv==1.0.0